requirements.txt
#Djago
Django>=2.1.3

#Django REST Framework
djangorestframework>=3.9.0

# Base
psycopg2>=2.7,<3.0
